+++ cmd/opentelemetry-prometheus-sidecar/main.go
@@ -24,18 +24,18 @@ import (
 	_ "net/http/pprof" // Comment this line to disable pprof endpoint.
 	"net/url"
 	"os"
-	"os/signal"
 	"path"
 	"runtime"
-	"syscall"
 	"time"
 
 	"github.com/go-kit/kit/log"
 	"github.com/go-kit/kit/log/level"
 	"github.com/lightstep/opentelemetry-prometheus-sidecar/config"
+	"github.com/lightstep/opentelemetry-prometheus-sidecar/health"
 	"github.com/lightstep/opentelemetry-prometheus-sidecar/metadata"
 	"github.com/lightstep/opentelemetry-prometheus-sidecar/otlp"
 	"github.com/lightstep/opentelemetry-prometheus-sidecar/retrieval"
+	"github.com/lightstep/opentelemetry-prometheus-sidecar/supervisor"
 	"github.com/lightstep/opentelemetry-prometheus-sidecar/tail"
 	"github.com/lightstep/opentelemetry-prometheus-sidecar/targets"
 	"github.com/lightstep/opentelemetry-prometheus-sidecar/telemetry"
@@ -46,11 +46,8 @@ import (
 	promconfig "github.com/prometheus/prometheus/config"
 	"github.com/prometheus/prometheus/pkg/labels"
 	"github.com/prometheus/prometheus/promql/parser"
-	"go.opentelemetry.io/contrib/instrumentation/net/http/otelhttp"
 	grpcMetadata "google.golang.org/grpc/metadata"
 )
-
-// TODO(jmacd): Define a path for healthchecks.
 
 // Note on metrics instrumentation relative ot the original OpenCensus
 // instrumentation of this code base:
@@ -64,6 +61,8 @@ import (
 // TODO(jmacd): Note that https://github.com/mwitkow/go-conntrack was removed, may
 // be useful after other matters are resolved.
 
+const supervisorEnv = "MAIN_SUPERVISOR"
+
 func main() {
 	if !Main() {
 		os.Exit(1)
@@ -71,123 +70,53 @@ func main() {
 }
 
 func Main() bool {
-
+	// Setup debugging helpers
 	if os.Getenv("DEBUG") != "" {
 		runtime.SetBlockProfileRate(20)
 		runtime.SetMutexProfileFraction(20)
 	}
 
+	// Configure a from flags and/or a config file.
 	cfg, metricRenames, staticMetadata, err := config.Configure(os.Args, ioutil.ReadFile)
 	if err != nil {
 		usage(err)
 		return false
 	}
 
-	vlevel := cfg.LogConfig.Verbose
-	if cfg.LogConfig.Level == "debug" {
-		vlevel++
-	}
+	// Should this process act as supervisor?  This uses the supervisor
+	// environment variable to avoid recursion.
+	isSupervisor := !cfg.DisableSupervisor && os.Getenv(supervisorEnv) == ""
 
-	if vlevel > 0 {
-		telemetry.SetVerboseLevel(vlevel)
-	}
+	// Configure logging and diagnostics.
+	logger := mainLogger(cfg)
 
-	var plc promlog.Config
-	plc.Level = &promlog.AllowedLevel{}
-	plc.Format = &promlog.AllowedFormat{}
-	plc.Level.Set(cfg.LogConfig.Level)
-	plc.Format.Set(cfg.LogConfig.Format)
-	logger := promlog.New(&plc)
-
-	level.Info(logger).Log(
-		"msg", "Starting OpenTelemetry Prometheus sidecar",
-		"version", version.Info(),
-		"build_context", version.BuildContext(),
-		"host_details", Uname(),
-		"fd_limits", FdLimits(),
-	)
-
-	if data, err := json.Marshal(cfg); err == nil {
-		level.Debug(logger).Log("config", string(data))
-	}
-
-	// We avoided using the kingpin support for URL flags because
-	// it leads to special cases merging configs and because URL
-	// parsing succeeds in cases w/o a scheme, needs to be
-	// validated anyway.
-	type namedURL struct {
-		name       string
-		value      string
-		allowEmpty bool
-	}
-	for _, pair := range []namedURL{
-		{"destination.endpoint", cfg.Destination.Endpoint, false},
-		{"diagnostics.endpoint", cfg.Diagnostics.Endpoint, true},
-		{"prometheus.endpoint", cfg.Prometheus.Endpoint, false},
-	} {
-		if pair.allowEmpty && pair.value == "" {
-			continue
-		}
-		if pair.value == "" {
-			level.Error(logger).Log("msg", "endpoint must be set", "name", pair.name)
-			return false
-		}
-		url, err := url.Parse(pair.value)
-		if err != nil {
-			level.Error(logger).Log("msg", "invalid endpoint", "name", pair.name, "endpoint", pair.value, "error", err)
-			return false
-		}
-
-		switch url.Scheme {
-		case "http", "https":
-			// Good!
-		default:
-			level.Error(logger).Log("msg", "endpoints must use http or https", "name", pair.name, "endpoint", pair.value)
-			return false
-		}
-	}
-
-	telemetry.StaticSetup(logger)
-
-	// Determine the diagnostics configuration.
 	diagConfig := cfg.Diagnostics
 
 	if diagConfig.Endpoint == "" && !cfg.DisableDiagnostics {
 		diagConfig = cfg.Destination
 	}
 
+	telemetry.StaticSetup(logger)
+
 	if diagConfig.Endpoint != "" {
-		endpoint, _ := url.Parse(diagConfig.Endpoint)
-		hostport := endpoint.Hostname()
-		if len(endpoint.Port()) > 0 {
-			hostport = net.JoinHostPort(hostport, endpoint.Port())
-		}
-		// Set a service.name resource if none is set.
-		const serviceNameKey = "service.name"
-		if _, ok := diagConfig.Attributes[serviceNameKey]; !ok {
-			diagConfig.Attributes[serviceNameKey] = "opentelemetry-prometheus-sidecar"
-		}
-
-		// TODO: Configure metric reporting interval, trace batching interval,
-		// currently there is no such setting.
-
-		defer telemetry.ConfigureOpentelemetry(
-			telemetry.WithLogger(logger),
-			telemetry.WithExporterEndpoint(hostport),
-			telemetry.WithExporterInsecure(endpoint.Scheme == "http"),
-			telemetry.WithHeaders(diagConfig.Headers),
-			telemetry.WithResourceAttributes(diagConfig.Attributes),
-			telemetry.WithExportTimeout(diagConfig.Timeout.Duration),
-			telemetry.WithMetricReportingPeriod(telemetry.DefaultReportingPeriod),
-		).Shutdown(context.Background())
+		tel := startTelemetry(diagConfig, isSupervisor, logger)
+		defer tel.Shutdown(context.Background())
 	}
 
-	// We instantiate a context here since the tailer is used by two other components.
-	// The context will be used in the lifecycle of prometheusReader further down.
-	ctx, cancel := context.WithCancel(context.Background())
+	// Start the supervisor.
+	if isSupervisor {
+		return startSupervisor(cfg, logger)
+	}
+
+	// Start the sidecar.  This context lasts the lifetime of the sidecar.
+	ctx, cancelMain := telemetry.ContextWithSIGTERM(logger)
+	defer cancelMain()
+
+	healthChecker := health.NewChecker()
 
 	httpClient := &http.Client{
-		Transport: otelhttp.NewTransport(http.DefaultTransport),
+		// Note: The Sidecar->Prometheus HTTP connection is not traced.
+		// Transport: otelhttp.NewTransport(http.DefaultTransport),
 	}
 
 	filters, err := parseFilters(logger, cfg.Filters)
@@ -233,6 +162,8 @@ func Main() bool {
 
 	outputURL, _ := url.Parse(cfg.Destination.Endpoint)
 
+	cfg.Destination.Headers[config.AgentKey] = config.AgentMainValue
+
 	var scf otlp.StorageClientFactory = &otlpClientFactory{
 		logger:   log.With(logger, "component", "storage"),
 		url:      outputURL,
@@ -265,64 +196,66 @@ func Main() bool {
 		cfg.Prometheus.MaxPointAge.Duration,
 	)
 
-	// Perform a test of the outbound connection before starting.
-	if err := selfTest(logger, scf, cfg.StartupTimeout.Duration); err != nil {
+	// Start the admin server.
+	go func() {
+		defer cancelMain()
+
+		server := newAdminServer(healthChecker, cfg.Admin, logger)
+
+		go func() {
+			level.Debug(logger).Log("msg", "starting admin server")
+			<-ctx.Done()
+			if err := server.Shutdown(context.Background()); err != nil {
+				level.Error(logger).Log("msg", "admin server shutdown", "err", err)
+			}
+		}()
+
+		if err := server.ListenAndServe(); err != http.ErrServerClosed {
+			level.Error(logger).Log("msg", "admin listener", "err", err)
+		}
+	}()
+
+	// Check the progress file, ensure we can write this file.
+	startOffset, err := readWriteStartOffset(cfg, logger)
+	if err != nil {
+		level.Error(logger).Log("msg", "cannot write progress file", "err", err)
+		return false
+	}
+
+	logStartup(cfg, logger)
+
+	// Test for Prometheus and Outbound dependencies before starting.
+	if err := selfTest(ctx, promURL, scf, cfg.StartupTimeout.Duration, logger); err != nil {
 		level.Error(logger).Log("msg", "selftest failed, not starting", "err", err)
 		return false
 	}
 
+	healthChecker.SetReady(true)
+
+	// Run three inter-depdendent components:
+	// (1) Target cache
+	// (2) Prometheus reader
+	// (3) Queue manager
+	// TODO: Replace this with x/sync/errgroup
 	var g run.Group
 	{
-		ctx, cancel := context.WithCancel(context.Background())
 		g.Add(func() error {
 			targetCache.Run(ctx)
 			return nil
 		}, func(error) {
-			cancel()
+			cancelMain()
 		})
 	}
 	{
-		term := make(chan os.Signal)
-		signal.Notify(term, os.Interrupt, syscall.SIGTERM)
-		cancel := make(chan struct{})
 		g.Add(
 			func() error {
-				// Don't forget to release the reloadReady channel so that waiting blocks can exit normally.
-				select {
-				case <-term:
-					level.Warn(logger).Log("msg", "Received SIGTERM, exiting gracefully...")
-				case <-cancel:
-					break
-				}
-				return nil
-			},
-			func(err error) {
-				close(cancel)
-			},
-		)
-	}
-	{
-		// We use the context we defined higher up instead of a local one like in the other actors.
-		// This is necessary since it's also used to manage the tailer's lifecycle, which the reader
-		// depends on to exit properly.
-		g.Add(
-			func() error {
-				startOffset, err := retrieval.ReadProgressFile(cfg.Prometheus.WAL)
-				if err != nil {
-					level.Warn(logger).Log("msg", "reading progress file failed", "err", err)
-					startOffset = 0
-				}
-				// Write the file again once to ensure we have write permission on startup.
-				if err := retrieval.SaveProgressFile(cfg.Prometheus.WAL, startOffset); err != nil {
-					return err
-				}
-				waitForPrometheus(ctx, logger, promURL)
-				// Sleep a fixed amount of time to allow the first scrapes to complete.
+				// Sleep to allow the first scrapes to complete.
 				select {
 				case <-time.After(cfg.StartupDelay.Duration):
 				case <-ctx.Done():
 					return nil
 				}
+
 				err = prometheusReader.Run(ctx, startOffset)
 				level.Info(logger).Log("msg", "Prometheus reader stopped")
 				return err
@@ -331,49 +264,26 @@ func Main() bool {
 				// Prometheus reader needs to be stopped before closing the TSDB
 				// so that it doesn't try to write samples to a closed storage.
 				level.Info(logger).Log("msg", "Stopping Prometheus reader...")
-				cancel()
+				cancelMain()
 			},
 		)
 	}
 	{
-		cancel := make(chan struct{})
+		stopCh := make(chan struct{})
 		g.Add(
 			func() error {
 				if err := queueManager.Start(); err != nil {
 					return err
 				}
 				level.Info(logger).Log("msg", "OpenTelemetry client started")
-				<-cancel
+				<-stopCh
 				return nil
 			},
 			func(err error) {
 				if err := queueManager.Stop(); err != nil {
 					level.Error(logger).Log("msg", "Error stopping OpenTelemetry writer", "err", err)
 				}
-				close(cancel)
-			},
-		)
-	}
-	{
-		cancel := make(chan struct{})
-		server := &http.Server{
-			Addr: cfg.Admin.ListenAddress,
-		}
-		g.Add(
-			func() error {
-				level.Info(logger).Log("msg", "Web server started")
-				err := server.ListenAndServe()
-				if err != http.ErrServerClosed {
-					return err
-				}
-				<-cancel
-				return nil
-			},
-			func(err error) {
-				if err := server.Shutdown(context.Background()); err != nil {
-					level.Error(logger).Log("msg", "Error stopping web server", "err", err)
-				}
-				close(cancel)
+				close(stopCh)
 			},
 		)
 	}
@@ -382,9 +292,8 @@ func Main() bool {
 		return false
 	}
 
-	// Note: It's unclear if this code path can execute, presently
-	// there's no intentionally graceful shutdown.
-	level.Info(logger).Log("msg", "shutting down")
+	// SIGTERM causes graceful shutdown.
+	level.Info(logger).Log("msg", "sidecar process exiting")
 	return true
 }
 
@@ -419,7 +328,7 @@ func (s *otlpClientFactory) Name() string {
 	return s.url.String()
 }
 
-func waitForPrometheus(ctx context.Context, logger log.Logger, promURL *url.URL) {
+func waitForPrometheus(ctx context.Context, logger log.Logger, promURL *url.URL) bool {
 	tick := time.NewTicker(3 * time.Second)
 	defer tick.Stop()
 
@@ -429,17 +338,18 @@ func waitForPrometheus(ctx context.Context, logger log.Logger, promURL *url.URL)
 	for {
 		select {
 		case <-ctx.Done():
-			return
+			return false
 		case <-tick.C:
 			resp, err := http.Get(u.String())
 			if err != nil {
-				level.Warn(logger).Log("msg", "query Prometheus readiness", "err", err)
+				level.Warn(logger).Log("msg", "Prometheus readiness check", "err", err)
 				continue
 			}
 			if resp.StatusCode/100 == 2 {
-				return
+				return true
 			}
-			level.Warn(logger).Log("msg", "Prometheus not ready", "status", resp.Status)
+
+			level.Warn(logger).Log("msg", "Prometheus is not ready", "status", resp.Status)
 		}
 	}
 }
@@ -458,24 +368,152 @@ func parseFilters(logger log.Logger, filters []string) ([][]*labels.Matcher, err
 	return matchers, nil
 }
 
-func selfTest(logger log.Logger, scf otlp.StorageClientFactory, timeout time.Duration) error {
+func selfTest(ctx context.Context, promURL *url.URL, scf otlp.StorageClientFactory, timeout time.Duration, logger log.Logger) error {
 	client := scf.New()
 
-	level.Info(logger).Log("msg", "starting selftest")
+	level.Debug(logger).Log("msg", "starting selftest")
 
-	ctx := context.Background()
 	ctx, cancel := context.WithTimeout(ctx, timeout)
 	defer cancel()
 
-	if err := client.Selftest(ctx); err != nil {
-		_ = client.Close()
-		return fmt.Errorf("could not send test opentelemetry.ExportMetricsServiceRequest request: %w", err)
+	// These tests are performed sequentially, to keep the logs simple.
+	// Note waitForPrometheus has no unrecoverable error conditions, so
+	// loops until success or the context is canceled.
+	if !waitForPrometheus(ctx, logger, promURL) {
+		return fmt.Errorf("Prometheus is not ready")
 	}
 
-	if err := client.Close(); err != nil {
-		return fmt.Errorf("error closing test client: %w", err)
+	// Outbound connection test.
+	{
+		if err := client.Selftest(ctx); err != nil {
+			_ = client.Close()
+			return fmt.Errorf("could not send test opentelemetry.ExportMetricsServiceRequest request: %w", err)
+		}
+
+		if err := client.Close(); err != nil {
+			return fmt.Errorf("error closing test client: %w", err)
+		}
 	}
 
-	level.Info(logger).Log("msg", "selftest was successful")
+	level.Debug(logger).Log("msg", "selftest was successful")
 	return nil
+}
+
+func mainLogger(cfg config.MainConfig) log.Logger {
+	vlevel := cfg.LogConfig.Verbose
+	if cfg.LogConfig.Level == "debug" {
+		vlevel++
+	}
+
+	if vlevel > 0 {
+		telemetry.SetVerboseLevel(vlevel)
+	}
+
+	var plc promlog.Config
+	plc.Level = &promlog.AllowedLevel{}
+	plc.Format = &promlog.AllowedFormat{}
+	plc.Level.Set(cfg.LogConfig.Level)
+	plc.Format.Set(cfg.LogConfig.Format)
+	return promlog.New(&plc)
+}
+
+func logStartup(cfg config.MainConfig, logger log.Logger) {
+	level.Info(logger).Log(
+		"msg", "Starting OpenTelemetry Prometheus sidecar",
+		"version", version.Info(),
+		"build_context", version.BuildContext(),
+		"host_details", Uname(),
+		"fd_limits", FdLimits(),
+	)
+
+	if data, err := json.Marshal(cfg); err == nil {
+		level.Debug(logger).Log("config", string(data))
+	}
+
+	if !cfg.DisableSupervisor {
+		level.Debug(logger).Log("msg", "running under supervisor")
+	}
+}
+
+func startSupervisor(cfg config.MainConfig, logger log.Logger) bool {
+	super := supervisor.New(supervisor.Config{
+		Logger: log.With(logger, "component", "supervisor"),
+		Admin:  cfg.Admin,
+	})
+
+	os.Setenv(supervisorEnv, "active")
+
+	return super.Run(os.Args)
+}
+
+func startTelemetry(diagConfig config.OTLPConfig, isSuper bool, logger log.Logger) *telemetry.Telemetry {
+	endpoint, _ := url.Parse(diagConfig.Endpoint)
+	hostport := endpoint.Hostname()
+	if len(endpoint.Port()) > 0 {
+		hostport = net.JoinHostPort(hostport, endpoint.Port())
+	}
+
+	insecure := endpoint.Scheme == "http"
+	metricsHostport := hostport
+	spanHostport := hostport
+
+	// Set a service.name resource if none is set.
+	const serviceNameKey = "service.name"
+
+	svcName := diagConfig.Attributes[serviceNameKey]
+	if svcName == "" {
+		svcName = "opentelemetry-prometheus-sidecar"
+	}
+
+	agentName := config.AgentSecondaryValue
+	if isSuper {
+		// Disable metrics in the supervisor
+		metricsHostport = ""
+		svcName = svcName + "-supervisor"
+		agentName = config.AgentSupervisorValue
+	} else {
+		// Disable spans in the main process
+		spanHostport = ""
+	}
+
+	diagConfig.Attributes[serviceNameKey] = svcName
+	diagConfig.Headers[config.AgentKey] = agentName
+
+	// TODO: Configure trace batching interval.
+
+	return telemetry.ConfigureOpentelemetry(
+		telemetry.WithLogger(logger),
+		telemetry.WithSpanExporterEndpoint(spanHostport),
+		telemetry.WithSpanExporterInsecure(insecure),
+		telemetry.WithMetricsExporterEndpoint(metricsHostport),
+		telemetry.WithMetricsExporterInsecure(insecure),
+		telemetry.WithHeaders(diagConfig.Headers),
+		telemetry.WithResourceAttributes(diagConfig.Attributes),
+		telemetry.WithExportTimeout(diagConfig.Timeout.Duration),
+		telemetry.WithMetricReportingPeriod(config.DefaultReportingPeriod),
+	)
+}
+
+func newAdminServer(hc *health.Checker, acfg config.AdminConfig, logger log.Logger) *http.Server {
+	mux := http.NewServeMux()
+	mux.Handle("/-/health", hc.Health())
+	mux.Handle("/-/ready", hc.Ready())
+	address := fmt.Sprint(acfg.ListenIP, ":", acfg.Port)
+	return &http.Server{
+		Addr:    address,
+		Handler: mux,
+	}
+}
+
+// readWriteStartOffset reads the last (approxiate) progress position and re-writes
+// the progress file, to ensure we have write permission on startup.
+func readWriteStartOffset(cfg config.MainConfig, logger log.Logger) (int, error) {
+	startOffset, err := retrieval.ReadProgressFile(cfg.Prometheus.WAL)
+	if err != nil {
+		level.Warn(logger).Log("msg", "reading progress file failed", "err", err)
+		startOffset = 0
+	}
+
+	err = retrieval.SaveProgressFile(cfg.Prometheus.WAL, startOffset)
+	return startOffset, err
 }