+++ otlp/client.go
@@ -29,7 +29,6 @@ import (
 	sidecar "github.com/lightstep/opentelemetry-prometheus-sidecar"
 	metricsService "github.com/lightstep/opentelemetry-prometheus-sidecar/internal/opentelemetry-proto-gen/collector/metrics/v1"
 	"github.com/lightstep/opentelemetry-prometheus-sidecar/telemetry"
-	"go.opentelemetry.io/contrib/instrumentation/google.golang.org/grpc/otelgrpc"
 	"go.opentelemetry.io/otel/metric"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/credentials"
@@ -72,6 +71,14 @@ var (
 		"points.exported",
 		metric.WithDescription("count of exported metric points"),
 	)
+	exportDuration = telemetry.NewTimer(
+		"otlp.export.duration",
+		"duration of the otlp.Export() call",
+	)
+	connectDuration = telemetry.NewTimer(
+		"otlp.connect.duration",
+		"duration of the grpc.Dial() call",
+	)
 )
 
 // Client allows reading and writing from/to a remote gRPC endpoint. The
@@ -102,28 +109,24 @@ func NewClient(conf *ClientConfig) *Client {
 	if logger == nil {
 		logger = log.NewNopLogger()
 	}
-	copyHeaders := map[string][]string{
-		telemetry.TelemetryReportingAgentKey: []string{telemetry.TelemetryReportingAgentMainValue},
-	}
-	for k, v := range conf.Headers {
-		copyHeaders[k] = append(copyHeaders[k], v...)
-	}
 	return &Client{
 		logger:           logger,
 		url:              conf.URL,
 		timeout:          conf.Timeout,
 		rootCertificates: conf.RootCertificates,
-		headers:          copyHeaders,
+		headers:          conf.Headers,
 	}
 }
 
 // getConnection will dial a new connection if one is not set.  When
 // dialing, this function uses its a new context and the same timeout
 // used for Store().
-func (c *Client) getConnection(ctx context.Context) (*grpc.ClientConn, error) {
+func (c *Client) getConnection(ctx context.Context) (_ *grpc.ClientConn, retErr error) {
 	if c.conn != nil {
 		return c.conn, nil
 	}
+
+	defer connectDuration.Start(ctx).Stop(&retErr)
 
 	ctx, cancel := context.WithTimeout(ctx, c.timeout)
 	defer cancel()
@@ -137,8 +140,10 @@ func (c *Client) getConnection(ctx context.Context) (*grpc.ClientConn, error) {
 
 	dopts := []grpc.DialOption{
 		grpc.WithBlock(), // Wait for the connection to be established before using it.
-		grpc.WithUnaryInterceptor(otelgrpc.UnaryClientInterceptor()),
 		grpc.WithDefaultServiceConfig(serviceConfig),
+
+		// Note: The Sidecar->OTel gRPC connection is not traced:
+		// grpc.WithUnaryInterceptor(otelgrpc.UnaryClientInterceptor()),
 	}
 	if useAuth {
 		var tcfg tls.Config
@@ -162,6 +167,11 @@ func (c *Client) getConnection(ctx context.Context) (*grpc.ClientConn, error) {
 				RootCAs:    certPool,
 			}
 		}
+		level.Debug(c.logger).Log(
+			"msg", "tls configured",
+			"server", c.url.Hostname(),
+			"root_certs", fmt.Sprint(c.rootCertificates),
+		)
 		dopts = append(dopts, grpc.WithTransportCredentials(credentials.NewTLS(&tcfg)))
 	} else {
 		dopts = append(dopts, grpc.WithInsecure())
@@ -250,7 +260,10 @@ func (c *Client) Store(req *metricsService.ExportMetricsServiceRequest) error {
 				ResourceMetrics: req.ResourceMetrics[begin:end],
 			}
 
-			if _, err := service.Export(c.grpcMetadata(ctx), req_copy); err != nil {
+			var err error
+			defer exportDuration.Start(ctx).Stop(&err)
+
+			if _, err = service.Export(c.grpcMetadata(ctx), req_copy); err != nil {
 				// TODO This happens too fast _after_ a healthy
 				// connection becomes unhealthy. Fix.
 				level.Debug(c.logger).Log(