+++ telemetry/telemetry.go
@@ -23,8 +23,8 @@ import (
 
 	"github.com/go-kit/kit/log"
 	"github.com/go-kit/kit/log/level"
+	"github.com/lightstep/opentelemetry-prometheus-sidecar/config"
 	"github.com/pkg/errors"
-	"github.com/prometheus/common/version"
 	hostMetrics "go.opentelemetry.io/contrib/instrumentation/host"
 	runtimeMetrics "go.opentelemetry.io/contrib/instrumentation/runtime"
 	"go.opentelemetry.io/contrib/propagators/b3"
@@ -40,24 +40,6 @@ import (
 	"google.golang.org/grpc/credentials"
 )
 
-const (
-	DefaultExportTimeout   = time.Second * 60
-	DefaultReportingPeriod = time.Second * 30
-
-	TelemetryReportingAgentKey = "telemetry-reporting-agent"
-)
-
-var (
-	TelemetryReportingAgentMainValue = fmt.Sprint(
-		"opentelemetry-prometheus-sidecar/",
-		version.Version,
-	)
-	TelemetryReportingAgentSelfValue = fmt.Sprint(
-		"opentelemetry-prometheus-sidecar-self/",
-		version.Version,
-	)
-)
-
 type (
 	Telemetry struct {
 		config        Config
@@ -67,32 +49,50 @@ type (
 	Option func(*Config)
 
 	Config struct {
-		ExporterEndpoint         string
-		ExporterEndpointInsecure bool
-		Propagators              []string
-		MetricReportingPeriod    time.Duration
-		ResourceAttributes       map[string]string
-		Headers                  map[string]string
-		ExportTimeout            time.Duration
-		resource                 *resource.Resource
-		logger                   log.Logger
+		SpanExporterEndpoint            string
+		SpanExporterEndpointInsecure    bool
+		MetricsExporterEndpoint         string
+		MetricsExporterEndpointInsecure bool
+
+		Propagators           []string
+		MetricReportingPeriod time.Duration
+		ResourceAttributes    map[string]string
+		Headers               map[string]string
+		ExportTimeout         time.Duration
+		resource              *resource.Resource
+		logger                log.Logger
 	}
 
 	setupFunc func() (start, stop func(context.Context) error, err error)
 )
 
-// WithExporterEndpoint configures the endpoint for sending metrics via OTLP
-func WithExporterEndpoint(url string) Option {
+// WithSpanExporterEndpoint configures the endpoint for sending spans via OTLP
+func WithSpanExporterEndpoint(url string) Option {
 	return func(c *Config) {
-		c.ExporterEndpoint = url
+		c.SpanExporterEndpoint = url
 	}
 }
 
-// WithExporterInsecure permits connecting to the
+// WithSpanExporterInsecure permits connecting to the
 // trace endpoint without a certificate
-func WithExporterInsecure(insecure bool) Option {
+func WithSpanExporterInsecure(insecure bool) Option {
 	return func(c *Config) {
-		c.ExporterEndpointInsecure = insecure
+		c.SpanExporterEndpointInsecure = insecure
+	}
+}
+
+// WithMetricsExporterEndpoint configures the endpoint for sending metricss via OTLP
+func WithMetricsExporterEndpoint(url string) Option {
+	return func(c *Config) {
+		c.MetricsExporterEndpoint = url
+	}
+}
+
+// WithMetricsExporterInsecure permits connecting to the
+// trace endpoint without a certificate
+func WithMetricsExporterInsecure(insecure bool) Option {
+	return func(c *Config) {
+		c.MetricsExporterEndpointInsecure = insecure
 	}
 }
 
@@ -161,10 +161,10 @@ func newConfig(opts ...Option) Config {
 	}
 
 	if c.ExportTimeout <= 0 {
-		c.ExportTimeout = DefaultExportTimeout
+		c.ExportTimeout = config.DefaultExportTimeout
 	}
 	if c.MetricReportingPeriod <= 0 {
-		c.MetricReportingPeriod = DefaultReportingPeriod
+		c.MetricReportingPeriod = config.DefaultReportingPeriod
 	}
 
 	var err error
@@ -214,25 +214,19 @@ func newExporter(endpoint string, insecure bool, headers map[string]string) *otl
 	if insecure {
 		secureOption = otlp.WithInsecure()
 	}
-	copyHeaders := map[string]string{
-		TelemetryReportingAgentKey: TelemetryReportingAgentSelfValue,
-	}
-	for k, v := range headers {
-		copyHeaders[k] = v
-	}
 	return otlp.NewUnstartedExporter(
 		secureOption,
 		otlp.WithAddress(endpoint),
-		otlp.WithHeaders(copyHeaders),
+		otlp.WithHeaders(headers),
 	)
 }
 
 func (c *Config) setupTracing() (start, stop func(ctx context.Context) error, err error) {
-	if c.ExporterEndpoint == "" {
+	if c.SpanExporterEndpoint == "" {
 		level.Debug(c.logger).Log("msg", "tracing is disabled: no endpoint set")
 		return nil, nil, nil
 	}
-	spanExporter := newExporter(c.ExporterEndpoint, c.ExporterEndpointInsecure, c.Headers)
+	spanExporter := newExporter(c.SpanExporterEndpoint, c.SpanExporterEndpointInsecure, c.Headers)
 
 	// TODO: Make a way to set the export timeout, there is
 	// apparently not such a thing for OTel-Go:
@@ -257,11 +251,11 @@ func (c *Config) setupTracing() (start, stop func(ctx context.Context) error, er
 }
 
 func (c *Config) setupMetrics() (start, stop func(ctx context.Context) error, err error) {
-	if c.ExporterEndpoint == "" {
+	if c.MetricsExporterEndpoint == "" {
 		level.Debug(c.logger).Log("msg", "metrics are disabled: no endpoint set")
 		return nil, nil, nil
 	}
-	metricExporter := newExporter(c.ExporterEndpoint, c.ExporterEndpointInsecure, c.Headers)
+	metricExporter := newExporter(c.MetricsExporterEndpoint, c.MetricsExporterEndpointInsecure, c.Headers)
 
 	pusher := controller.New(
 		processor.New(
@@ -305,9 +299,8 @@ func ConfigureOpentelemetry(opts ...Option) *Telemetry {
 		config: newConfig(opts...),
 	}
 
-	level.Debug(tel.config.logger).Log("msg", "debug logging enabled")
 	s, _ := json.MarshalIndent(tel.config, "", "\t")
-	level.Debug(tel.config.logger).Log("configuration", string(s))
+	level.Debug(tel.config.logger).Log("msg", "telemetry enabled", "cfg", string(s))
 
 	var startFuncs []func(context.Context) error